Add a mutual-exclusion Semaphore to PropertyManager
In Dolphin this is internal to the WeakIdentityDictionary, but Pharo weak collections don't have this protection built-in. Prevents property table corruption when multiple processes use properties with different objects. Note that multiple processes accessing properties of the *same* object can still cause problems, but this is the case in Dolphin as well.